Minor change nint Crash with "Object used after it is disposed" in the GetFlipped() method. This occurs after a while of redrawing the image, especially after adding 3 or 4 images. Investigate

+++ BNR_Cocoa_Book/DrawingFun/DrawingFun/MainWindowController.cs
@@ -36,7 +36,7 @@ namespace DrawingFun
 		{
 			NSOpenPanel panel = NSOpenPanel.OpenPanel;
 			panel.AllowedFileTypes = NSImage.ImageFileTypes;
-			panel.BeginSheet(stretchView.Window, (int result) => {
+			panel.BeginSheet(stretchView.Window, (nint result) => {
 				if (result == 1) {
 					StretchImage image = new StretchImage(panel.Url);
 					stretchView.Image = image;
